requirements.txt
plotly.express
numpy
pywebio
matplotlib
seaborn
typing
pandas
